LastfmStreaming: Delete the Loved stations from the database
The Loved station was discontinued by Last.fm, so it was removed from
the code in a previous commit. Having a station of that type still
present in the database causes an exception to be logged on startup.

Remove any existing stations of that type from the LastfmStations table,
as those stations would be unusable and not visible anyway. Fixes
bgo#639032.