SoundMenu: Implement the new registration process
The sound menu included in Ubuntu Natty (indicator-sound > 0.5.8)
now automatically detects MPRIS applications appearing on the bus. So we
don't have to register explicitly anymore.

Unregistering is done by adding ourselves in a black-list of players
which won't appear in the sound menu. This should be done through
GSettings, but indicator-sound kindly provides us with a d-bus method
for that.

This means that the build dependency on indicate-sharp is now optional
and can be substituted by a runtime dependency on indicator-sound >
0.5.8.

For compatibility with Ubuntu Maverick, we still do the explicit
registration with indicate-sharp. This will just be ignored by
indicator-sound provided on Natty.

This is all hidden away in a new SoundMenuProxy class, to keep things
reasonably sane.