windows: Fix "Launch Banshee" checkbox in installer (bgo#641989)
In order to run a file that was just installed, we need to reference it
in the FileKey attribute, using its id.

To ba able to do that, we have to disable the automatic id generation
done by the heat tool, so that the id will now be the file name. We can
only do that for the bin directory, because otherwise we get id
conflicts.